+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ### Bug Fixes
 
 -   Fixed missing comments on `@enum` style enum members defined in declaration files, #1880.
+-   Fixed `--validation.notDocumented` warnings for functions/methods, #1895.
 -   Search results will no longer include random items when the search bar is empty, #1881.
+-   Comments on overloaded constructors will now be detected in the same way that overloaded functions/methods are.
 
 ### Thanks!